feat: 게시글 상세 본문 푸터와 플로팅 푸터 상태 공유 구현
